chore(sveltekit): refactor some common server-side code
Handle functions and utils.ts

There is more left to refactor as well
