+++ packages/sveltekit/src/server/serverRoute.ts
@@ -1,7 +1,7 @@
 import { addNonEnumerableProperty } from '@sentry/core';
 import { SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN, SEMANTIC_ATTRIBUTE_SENTRY_SOURCE, startSpan } from '@sentry/node';
 import type { RequestEvent } from '@sveltejs/kit';
-import { flushIfServerless, sendErrorToSentry } from './utils';
+import { flushIfServerless, sendErrorToSentry } from '../server-common/utils';
 
 type PatchedServerRouteEvent = RequestEvent & { __sentry_wrapped__?: boolean };
